fix error: adding element to panel
