+++ components/CatalogContainer.vue
@@ -11,13 +11,7 @@
     let unsubscribeFromAddElement; // Store the unsubscribe function
 
     function addNewElementToDisplay(event) {
-        let fixedHeight = 200;
-        let name = event.target.alt;
-        let src = event.target.src;
-        let width = (fixedHeight * event.target.naturalWidth) / event.target.naturalHeight;
-        let newAsset = new Asset(src);
-        let tempEl = new ElementDS(width, fixedHeight, name, src, newAsset);
-        comicStore.bus.emit('add-element', tempEl);
+        comicStore.bus.emit('add-element', event);
     }
 </script>
